backend/requirements.txt
fastapi
uvicorn
numpy
psycopg2-binary
python-jose
passlib
python-dotenv
bcrypt